\name{GetNetMHCIIpan}
\alias{GetNetMHCIIpan}
\title{Helps downloading and installing NetMHCIIpan for MHC-II binding predictions}
\usage{
GetNetMHCIIpan(version_number = "4.1")
}
\arguments{
  \item{version_number}{the version of NetMHCIIpan to be downloaded.}
}
\description{The function leads the user through the downloading and installing process on NetMHCIIpan. The function works only in a Linux-based R environment.}