\name{GetNetMHCpan}
\alias{GetNetMHCpan}
\title{Helps downloading and installing NetMHCpan for MHC-I binding predictions}
\usage{
GetNetMHCpan(version_number = "4.1")
}
\arguments{
  \item{version_number}{the version of NetMHCpan to be downloaded.}
}
\description{The function leads the user through the downloading and installing process on NetMHCpan. The function works only in a Linux-based R environment.}